more files for rudimentary group-leaderboard

+++ src/main/java/co/grandcircus/challengeaccepted/projectioninterfaces/LeaderboardRow.java
@@ -2,7 +2,7 @@ package co.grandcircus.challengeaccepted.projectioninterfaces;
 
 public interface LeaderboardRow {
 	
-	Integer getUserId();
+	Long getUserId();
 	String getFirstName();
 	String getLastName();
 	Integer getCompleted();
